first version of templates for metadata
This version introduces the ability to customize the commit message
and author for an import.

The argument receives a list of functions that are executed
sequentially. The reason is that the strategy we are following
is to have several independent built-in functions thad add
information or transform the metadata. This will avoid making
users to write their own functions. We probably will have some
built-in because I think I can do some partial functions that
cannot be done in pure Skylark. IOW being able to have a
parametrizable function that returns a function. Like:

   message_transformers = [add_header("Copybara project")]

It turns out that Skylark doesn't support Iterable, so we have to
use SkylarkList instead. But for now is fine since even if we load
a massive repo AND we load all the commits, we would have commits
in the order of thousands.

There are stil many things left. For example:

   - Dealing propertly with labels.

   - Remove hardcoded message generation and replace them
     by builtin functions.

   - Add git.squash_notes like builtin to create a nicer
     message with short sha1 ids.

   - A function for removing and adding a header (Internal).

   - Get rid of OriginalAuthor and merge Change and SkylarkChange

Change-Id: I28956d933cfca50ca89e17f077786e5a3d5da275