Allows the resolution of the git binary using env vars.
The Git env var GIT_EXEC_PATH determines where Git looks for its
sub-programs, but also the regular git binaries (git,
git-upload-pack, etc) are duplicated in that path.

If the env var is not set, then we will execute "git", that
it will be resolved in the path as usual.

This is needed for environments where Git is not installed by
default and we need to install it and invoke it from the same
Java process (like the service). Adding the installation bin/
folder to the path does not work as the current process
environment is read-only.

Change-Id: Ifccfbedf2cf9a2f99a26ed3f01fe86ec5819736f