Switch RE engine from java.util.regex to re2j
This requires some tricks because re2j does not support named groups,
which means we must use group indexes. We keep track of the number of
groups in each group in order to properly build the replacement regex
(note that the user can include '()' in regex_groups, which makes the
number of actual groups per user-defined "regex_group" unpredictable.

This also adds a test to make sure that the regex group is not
repeated in the before or after template, since this was never
supported before and will not be possible with re2j, which does not
support backreferences.

Change-Id: I09b037c07a15990c59c391ea092f9570bdd88776