Remove OriginalAuthor class
This change removes OriginalAuthor and relies on the origins to
construct the author properly (calling to authoring). This allows
that the changes that are returned by the origins can be used
in Skylark direclty without needing to keep a copy of the object.

I also think it simplifies the code. Specially for the internal
plans.

Internal repo will be broken after this change. I'll fix on import.

Change-Id: Ie0d9ff00c41dba584819171ba8679ab6c53b0c3e