Refactor Destination API: move destination_files to Writer field
This moves the getPreviousRef method to the Writer as well. In this
way, the Destination implementations will not need their own 'paths'
argument (this pattern occurs in the internal repo) and can repurpose
the field originally meant for specifying exclusions to also specify
what paths to import into for monolithic repos.

I want to do this for Origin as well in a separate commit, since it is
a non-trivial refactor.

Change-Id: I75bcff073a6cfca1bd667fddbb83b1e4683048e1