Refactor Origin to accept origin_files glob
This is useful if an Origin implementation wants to use the
origin_files glob to see which parts of the depot are being exported,
for the purposes of optimizing queries.

Because origin_files is a workflow setting which is not known when the
origin is constructed, we introduce Origin.Reader, which is an Origin
that reads from a particular set of paths. Authoring is also required
to construct Origin.Reader, since many methods of Reader would require
authoring, and Authoring is also a workflow setting.

Change-Id: I1eb2376aefff4c6d6a08ac26cefc65d42327edee