Add support for manipulating labels in transforms
This change allows to manipulate labels on the constructed
message. While changes has a convenient method that returns
the map of labels. This code acts on the generated message
(That could be different from the multiple change messages
that are imported).

I am thinking that I'll probably remove that map and use the
new functions (modified to work for Change). But I prefer to
do it in a different change.

Change-Id: I77946fb2ac20510969c9375ba51c9b16f75f0272