Log the path and content of the main config file
This change logs the path and content of the main config file. There
is a common pattern in our users that are experimenting with
Copybara but they haven't already checked in any config file. Given
that the typical config file size should be relatively small, lets
log it so that we can repro their issues.

Change-Id: I84a2e85cbd82650985dff831016692a003c6caa3