Be able to add a header to changes summary
This transformer allows to add a header to a change summary. The
text can contain references to the current message labels or
labels on the original changes messages.

The way of referencing those labesl would be ${LABEL}.

We might generalize this method to a general template system, but
adding a header is what we need for now.

Change-Id: I783b68da84162008845fbdcfc400c0cd711a6728