Allow repeated groups in replace
This change allows to use repeated groups in replace regex. Note that
this is done without backtracking, so there are some limitations when
there are multiple matches, since it will pick the first one and if
it fails it will not try again.

The purpose of this mode is to be able to replace repeated strings
when there is a clear separator, like paths or package names. For
example:

  - foo.google.google

  - google/foo/google

Change-Id: I68303711ea286ba79021455b817f1fc3c2e837d6