User transforms: Be able to use globs and paths
This is the first change for having user defined transforms that can
access the checkout files, manipulate them using skylark or native
functions and write them back.

In this change we introduce the ability to glob the checkout dir and
be able to create derivated paths of those files (path.parent,
path.resolve, etc.).

The testing part is tricky since we cannot do too much in the
transform for now. So the change message is used for the output.

Change-Id: I3f7deb887437898c9f869d45a53ed2a7c2707931