Generalize Workflow to Migration
This change generalizes the the Workflow type to Migrations. Migrations
represents file moves between one-many sources to one-many destinations
that don't necessarily require transformations, authoring, etc.

This is needed to be able to have git.mirror() like migration or other
internal use cases.

We will probably need some sort of introspection for migrations in order
to know from/to where code is being moved. But that only requires url
like fields and that is not currently exposed in Origin/Destination.

Change-Id: Ia7578189fdca61770624c5446e6e9f636ed4133a