Introduce [command] argument.
New usage:
copybara [COMMAND] CONFIG_PATH [WORKFLOW_NAME [SOURCE_REF]]

The default command is "migrate", and the other possible
commands for now are:
- validate: invokes the validation
- info: reads the last migrated reference

In this way we don't need to introduce a new boolean
flag every time we need a new command for the tool. The
--validate flag has been removed.

This change also includes integration tests for the
validation and tests for command line argument parsing.

Change-Id: Ia7a575b6b7194662419e67638a5370ba2f46a457