git mirror: refspec reference converter
This change allows Refspec to translate origin refs to the destination
one.

This will be used to ls-remote both the origin and destinations and
compute a diff of the changed references so that we can avoid the
origin fetch if nothing changed. While this operation will add some
latency (2-3 seconds), it will be more efficent for environments where
we the origin repo cache is not maintained.

Change-Id: I42e8a6efd166bace95ec2118b0eac1b64063d3d7