Introduces a capturing console.
This is a console that captures the messages (by level) and
delegates to another console. By default it captures all the
messages, but it can also filter only specific message types.

The main usage of this right now is internal, but it's general
enough to make it available here. Also we are taking advantage
of existing test classes promoting them to production ones.

Change-Id: I449344578b598ca96c6d8ad7fc213a1910b0eac0