Reduce 20x the time for commit in git.destination
This change reduces the time of adding back the excluded files by
20x. Previously this code was taking a significant amount of time
for internal medium size repos.

The strategy is that we visit/list all the files in the checkout
dir (fast operation) and then we batch the 'git add' operations
up to a conservative size (6K of params) to avoid args too long
error.

The major issue of this solution is a possible memory comsumption,
since we load all the file names in memory. But I don't expect this
to be a problem since we would need a million of excluded files to be
a problem (Well maybe not that many, but you get the idea). At that
point we could implement an streaming strategy. But this is good
enough.

As part of this change I've moved the 'production' usages of
simpleCommand to a specialized addX method. There are still tons
of usages in test, and we have to make a more general method. But
not in this change.

Change-Id: I82de640d4048f1fbe440166c7060053a814b007e