Fix submodules support for git.origin()
Our previous version of submodules was implemented incorrectly, since
the submodule commit wouldn't be constant for a given parent commit.
Instead, we where blindly checking out head master (or the
corresponding branch).

Submodules don't work in that way. Given a commit in the parent git
repo, a submodule in that repo will point to an specific constant
reference of the submodule repo.

As an added benefit we don't have to check for cycles anymore since
infinite recursive SHA-1 are imposible.

We still have to fix '.' branches. They should match the parent
branch. The current implementation will work for branches (but
not for non-standard refs like refs/foo/bar).

Change-Id: I318364513e002b9e9a87a16640b29ec61c1ba443