Add --force flag for migrations
This is the first change of a plan of making Copybara safer to run
by default.

This will prevent users from accidentally migrating code backward
in time and automatized versions of Copybara reliable.

The next changes will add a validation for any workflow that we
can read lastRev in the destination and check with the origin
that it is an ancestor of the current imported change.

Currently in git.mirror we don't force. This will allow us to force
if the user requests it.

Change-Id: I4f11fafa3975b01b55c5014052b3510b242a613a