Remove --first-migration flag and use --force instead
There is currently no reason for not just using --force. This makes
it easier for our users to understand the new semmantics that we want
to have: If we know the state of the destination/origin we know
that it is safe to run the migration. Otherwise ask for --force.

Change-Id: If48d4a4588b8a84a210c7e088d248a1399ee75f0