Don't print folder.destination path until used
Config files can contain multiple workflows. We were printing at
parsing time to the console instead of printing the path only
when the workflow is really invoked.

This was very confusing as I run a workflow that didn't use
folder.destination() but still printed that a folder would be
used for the output.

Change-Id: I93b1cb56fdc718806595aea6e38924d6e5af278c