Make squash workflow safe by default
This change makes squash safe by default. We get the latest migrated
change in the destination and we check that it is an ancestor of the
current one.

The check is skipped if the origin/destination don't support history
(basically folder.* and some test origin/destinations). We might
change the new interface methods in the future but this is good enough
for now.

As part of this change I also fix an issue where if --force was used
we would fail in git.destination if the fetch branch existed. This
was introduced in the transition from --first-commit to --force.

Change-Id: I872950208bf01952746dd54da8211b82dc67eebc