Clear unstaged files for git.destination rebase
When git.destination does a rebase (CHANGE_REQUEST workflow), it
requires that the work-tree doesn't contain any unstaged files.
The reason is that ther might be a conflict and rebase puts in
the work-tree the contents of files and the conflict files.

The current git.destination logic is too complicated (we do a
scratch clone and then manipulate it with an alternate work-tree).
We should make this simpler and use the repo cache. But nothing
that we didn't know already :).

Change-Id: Iba40f529ca5644f4ee4f4a62c6e4f31f2c76c365