Iterative limit the number of changes to migrate
This change introduces the ability to limit the number of pending
changes to migrate in iterative mode.

The main usage of this is to be able to run just one migration
at a time (And run some process in the middle). But also to be
able to run migrations in batches.

Change-Id: Ib752dd9d618314b1aea2e901502e69eea0bba343