Issue an error if destination_files is too narrow
This causes an error to occur if destination_files does not encompass
every file being written to the destination. This is done in Workflow
so that it is automatically applied to all Destination types.

The ValidateDestinationFilesVisitor explains why this is helpful.

This rolls-forward commit 767c0292d622bf46de293ebbbe480dc3ea66343c
which had been rolled back because an internal user was still relying
on the quirk, but they have since fixed it.

Change-Id: Ib1dcfed07c17102a6124ce62f103ac45fd2c65d8