Allow ReadablePathMatcher to be usable on Windows
Trying to resolve a glob on Windows fails because the glob character "*"
is an invalid character in Windows files. This happens to work on
POSIX file systems because the glob character is valid there.
Path.resolve(String) really only expects to be receive normalizable
paths, but Globs are not normalizable since they represent a potential
multitude of files.

Instead of using Path.resolve(String) just concatenate the normalized
root path with the current filesystem's separator and then add the glob
expression on the end.

BUG=34421807
Change-Id: Idd61df509413556c53fd6924833cbc9884b29e04