Refactor of the configuration loading and parsing
This is some preparation work to be able to load configurations from
origins (something like --config-read-from-origin), required by some
clients.

This refactor moves the config file loading and parsing into a class
that can be extended by the internal version of Copybara. This also
removes the dependency on SkylarkParser in some parts of the code.

As part of this change, I've also renamed ConfigSupplier to
ModuleSupplier, as it was a misleading name.

Other future implementations of ConfigLoader might memoize the
configuration or load it from a specific changelist.

BUG=33381288

Change-Id: Id0d03e71a12fb7ffe8d60d0a25e86b1ce9dbe830