Allow ReadablePathMatcher to usable on Windows, part 2
On Windows the path separator will be a backslash and globs in copybara
are specified with the forward-slash path separator. The forward slashes
must be translated to the current filesystem's separator or globs
refering to subdirectories will not work.

For instance in directory "c:\tmp" applying glob "foo/**" will check
that "c:\tmp\foo\bar" from the filesystem walk to the glob
"c:\tmp\foo/**" and the forward slash will not match the backslash.

BUG=34421807
Change-Id: I918e724fe5e0b023981dc225555cced0922d96e5