DummyRevision can hold context references and labels
This change makes it possible for DummyRevision and DummyOrigin to
hold context references and labels so that we can use in tests
in next changes.

This will allow us to run a workflow for 'HEAD' and check that
we can access to the unresolved revision (the reference) and
in the future the labels.

Change-Id: I64ab13a831a8c9dad8a3f5610ba4eef2e2b7300d