Create migration identity for workflows
This change exposes a stable identifier for workflows that use context
references ('master' instead of SHA-1 for example) that keeps being
constant as long as neither the top level config file location
or the workflow name change.

For computing the config path, the ConfigFile tries to relativize
using the root path. This allows to move the folders containing
the config as long as the relative location to the root doesn't
change.

As part of this change we make the identity available to the
destination at the time of creating the writer. This would
allow destinations to find the entity to reuse sooner and
compute getPreviousRef correctly.

PiperOrigin-RevId: 148468052
Change-Id: I830e73e7563c2d7220b7705c707e9aa3860bccb6