Use last affected change instead of HEAD in SQUASH
Squash should work in the same way ITERATIVE works. Using
the most recent change as revision instead of arbitrary
head version. This is particularly bad on repositories
where we are only interested in a small set of the
directories (origin_files glob).

While we are still not fully correct, since there could
be some files excluded in origin_files and use that file
change, it is better than before. And opens the door too
fix it in the specific origins.

As part of this change I also changed the semantic when
lastRev is not found. Now we expose the whole history
to metadata.squash_notes. I added a flag to disable it.

BUG=35258491

PiperOrigin-RevId: 149347882
Change-Id: Iaa28b61c5f5e72986fad9277f4b24126a68f0153