Allow users to set the local git repo in destination
This change allow to specify a directory to use as the local git repo
used for git.destination. This allows to introspect or modify the
state after Copybara finished. Optionally the push to the remote
repo can be skipped too.

This should only be used for git features that Copybara doesn't
implement and is subject to change if we add those features.

The code is not extremely complex but we had to do some things:

  - We add as 'origin' remote the destination url. We also configure
    HEAD to push to the destination push ref. This allows the users
    of these repos to simply do 'git push' when done instead of
    knowing the destination url and ref and duplicating that
    information.

  - We cleanup the local worktree since we use a non-bare repo and
    otherwise the users would be confused because of local changes.

  - We switch to a ref with the same name as the destination. So if
    the git.destinaton push ref is 'master' the local repo would be
    pointing to 'master'. I had to use ls-remote to the repo to
    correctly resolve if it is a branch or tag.

  - If we skip the push we print the local repository location

Finally, this only works if the same Destination.Writer is used.
There is an internal instance for reading non-head configs that
would not work with this. But we don't allow the flag there.

BUG=36103055

Change-Id: I2e0585d732a37e4e77900347d8c60a1f5614b083