Move migration identity logic to migration call
We only create the destination writers at the beginning of the
migration. It doesn't make sense to compute the identity at that
time since it can change for ITERATIVE mode for example.

Furthermore, the original reason was to be able ot use it in
getLastRev. but getLastRev only reads submitted revisions.
We might want to have this for git.github_pr_destination but
lets fix it when we have a better idea of what we need then.

PiperOrigin-RevId: 150127867
Change-Id: Iafed787e2a240fe1b9a371218e5f53e469ee0888