Allow caching file visitations between transforms
This change makes it possible that if we have two consecutive
transforms where the first one uses the the TreeState API
for doing lookups and notify of modified files, the next one
will get access to a cached version of the TreeState.

Furthermore, if they use the same Glob, even the matching
filtering will be cached.

PiperOrigin-RevId: 150463557
Change-Id: I29d297d006521cce79177bc1b8ef332c306f4cc0