Add an extensible profiler
This change introduces an extensible profiler to Copybara. The
profiler supports hierarchical recording of stats using a namespaces
based description.

We use InheritableThreadLocal in order to keep track of the parent
task when running children in different threads.

This profiler should be relatively cheap. It only pays the penalty
if the child thread tries to record a metric. And even then we
create a deque of size 2.

The mechanism for extending the Profiler is to add callbacks that
will receive start/end events.

Change-Id: I398efcd1df9e4c7ab23aef1c925bc8cef1cc1b48