Add profiling to most important workflow components
Introduced also profiler utilities in generalOptions for dealing
with repositories. This is useful when a piece of code retrieves
some object so that we don't have to split the variable
declaration from the assignment. It should be used in the origin/
destinations. If it turns out that it is not that useful I will
inline/remove it from Workflow.

Consolidated Workflow generalOptions parameters in one argument.

Also move profiler to its own BUILD package

Change-Id: Ia22e80564a5654f67bdfad7dbbb2172d962f1dcc