Iterative skips excluded changes
This change allows to skip changes that only contain excluded files.
This is both a performance optimization (for large set of internal
only changes this would run a lot of noop migrations) but also for
correcteness, since metadata validators might fail for changes
that are only mean to be internal.

For this to happen the origin thas to provide a Change object with
the affected files set. Otherwise it is migrated since we cannot
know if the change is going to be a noop.

We also run the migrations that even it they don't touch any
included files it might touch Copybara configuration files. This
is for the cases where we store the config in the origin. The
mechanism doesn't differentiate between config stored in origin
or other storage, but false positives would be no harmful
(For example we run all the changes unconditionally now).

We use endsWith instead of equals for config equality because the
origin folder might a parent folder of the configuration root.

Fixed also some style errors.

BUG=36361543

PiperOrigin-RevId: 150960366
Change-Id: I7815c1aaab34391f5dffa5a299226341d7b09ffe