Add a patch 'series' parameter to the patch.apply module.
Patch management tools like QUILT [https://linux.die.net/man/1/quilt]
use a separate directory and a file listing the patches to apply. In
the most common scenario QUILT will store the patches in the './patches'
directory while maintaining an ordered sequence of the patch file names
in he './patches/series' file.

The CL adds a Skylark parameter 'series' to the PatchModule.APPLY
function. The patches are read from the series file at the time when
the configuration is parsed and the patch file names added to the other
patches.

PiperOrigin-RevId: 151704278
Change-Id: Ic05fe9d31d887f0602f7155044eb4e5036ef852d